// yourswami/src/lib/api-sports-types.ts
export interface GameTeam {
    id: number;
    name: string;
    logo: string;
}

export interface GameScoreInnings {
    [key: string]: number | null;
}

export interface GameScore {
    hits: number;
    errors: number;
    innings: GameScoreInnings;
    total: number | null;
}

export interface GameStatus {
    long: string;
    short: string;
}

export interface Game {
    id: number;
    date: string;
    time: string;
    timestamp: number;
    timezone: string;
    week: string | null;
    status: GameStatus;
    country: { id: number; name: string; code: string; flag: string };
    league: { id: number; name: string; type: string; logo: string; season: number };
    teams: {
        home: GameTeam;
        away: GameTeam;
    };
    scores: {
        home: GameScore;
        away: GameScore;
    };
}